Update multiplier_rounder.vhd
Implemented unsigned, signed and STD_LOGIC_VECTOR multiplication functions. The unsigned and STD_LOGIC_VECTOR got to work in a very small time. But the signed one was very tricky. But in this version I have verified that the signed multiplication now also works (even though the results are odd, they are correct).