Update BPF + sinewave gen
Corrected for -6dB resonance frequency in BPF filter computation by multiplying gain by 2 default.
Changed sinewave generic desired_freq variable to an input for doing tests with different frequencies, so the frequency spectrum could be seen.
New results of the frequency spectrum response for a BPF of 400Hz have been made and varying the gain so the signal will be clipped is also simulated corrreclty. These PDF files are also in this commit.