Updated BPF filter VHDL
Implemented Fixed Point in the BPF VHDL code. Now the code is synthesizable. Sadly the included BPF in the system results in a tremendous amount of noise, the audio is slightly hearable. But the noise is unbearable.
Replaced the gain computation with a lookup table that contains all the gain values for the different dB's from the param input.
Did the frequency test and gain test again. Simulation results can be seen in the .pdf from this commit.