Update BPF VHDL
Made a new schematic file in quartus to test the system with the BPF. Tried to fix it by making the gain lookup table a constant but it still just gives out noise a lot. The music can be heard very little.
So skip the BPF and focus on the report.