Update Audio-DSP
Implemented volume range from -87dB to 40dB. Removed clock divider and made MCLK for audio_codec with PLL. This seems to have fixed a bug where sometimes the left audio channel was not working.
The working volume controls from -87dB to 40dB are saved in the .sof files from this commit.
Also made a text file with the C code to generate the lookup table for the gains.